:recycle: 랜덤 인터뷰 페이지 SettingForm 컴포넌트 추상화
